unify kl-image renderers
