+++ src/components/Matching/MatchSvg/ImageSvgChoice.vue
@@ -21,11 +21,9 @@ export default {
     const url = ref(null)
     onMounted(async () => {
       const imageUrl = await Agent.download(props.imageId).url()
-      console.log('imageUrl', imageUrl) 
       url.value = imageUrl
     })
     return { url }
   }
 }
 </script>
-
