--- src/components/Rearrange/kl-image.vue
@@ -1,20 +1,0 @@
-<template>
-    <img :src="imageUrl" />
-</template>
-
-<script setup>
-    import { ref, watch } from 'vue'
-
-    const props = defineProps({ id: String })
-    const imageUrl = ref(null)
-
-    watch(() => props.id, getURL)
-
-    getURL()
-
-    async function getURL() {
-        console.log('FIRING!', props)
-        const url = await Agent.download(props.id).url()
-        imageUrl.value = url
-    }
-</script>
