+++ src/helpers/questionTypes.js
@@ -8,6 +8,8 @@ import SequenceCustomizer from '../components/Sequence/Customizer.vue'
 import SequencePlayer from '../components/Sequence/Player.vue'
 import SelectionOfWordsCustomizer from '../components/SelectionOfWords/Customizer.vue'
 import SelectionOfWordsPlayer from '../components/SelectionOfWords/Player.vue'
+import RowSelectionOfWordsCustomizer from '../components/RowSelectionOfWords/Customizer.vue'
+import RowSelectionOfWordsPlayer from '../components/RowSelectionOfWords/Player.vue'
 
 const questionTypes = {
 	'application/json;type=matching' : {
@@ -73,7 +75,27 @@ const questionTypes = {
 			choices: [],
 			answer: []
 		}
-	}
+	},
+	'application/json;type=selection-of-words-row' : {
+		description: 'Selection of Words Row',
+		components: {
+			player: RowSelectionOfWordsPlayer,
+			customizer: RowSelectionOfWordsCustomizer
+		},
+		newItemSchema: {
+			audioId: null,
+			choices: [
+				{	
+					content: 'Set Choice 1', // display string or uuid
+					correct: false
+				},
+				{	
+					content: 'Set Choice 2', // display string or uuid 
+					correct: true
+				}
+			]
+		}
+	},
 }
 
 let sequenceImportableTypes = Object.keys(questionTypes)
